backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
httpx>=0.24.0
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.43.0
pytrends==4.9.2
Pillow>=11.0.0
google-generativeai==0.5.0
apscheduler==3.10.4
python-multipart==0.0.9
websockets==12.0
amazon-paapi5>=1.1.0
boto3>=1.34.0
supabase>=2.0.0
